tmux: re-enable active pane highlighting
now that we use a color profile that inverts
the meaning of black and white,
we can use absolute colors (in terms of black
and white) which are then mapped by the color
profile to the actual colors.
This way, although tmux uses white and brightwhite,
we get a dark and very dark background in the dark
terminal profile.