lvim: set stack jumplist option
This makes the jump list behave like a stack
(or browser history, for that matter).

This means that when jumping from A to B to C,
then back to B, and then to D, jumping back
will jump to B and A, ignoring C because it was
already jumped back from (popped from the stack).